+++ python/CHANGELOG.rst
@@ -47,9 +47,21 @@
 - Add support for missing data to ``write_vcf``, and add the ``isolated_as_missing``
   argument. (:user:`jeromekelleher`, :pr:`2329`, :issue:`447`).
 
-- Add ``Tree.num_children_array`` and ``Tree.num_children``. Returns the counts of 
-  the number of child nodes for each or a single node in the tree respectively. 
+- Add ``Tree.num_children_array`` and ``Tree.num_children``. Returns the counts of
+  the number of child nodes for each or a single node in the tree respectively.
   (:user:`GertjanBisschop`, :issue:`2318`, :issue:`2319`, :pr:`2332`)
+
+- Add ``Tree.path_length``.
+  (:user:`jeremyguez`, :issue:`2249`, :pr:`2259`).
+
+- Add B1 tree balance index.
+  (:user:`jeremyguez`, :user:`jeromekelleher`, :issue:`2251`, :pr:`2281`, :pr:`2346`).
+
+- Add Sackin tree imbalance index.
+  (:user:`jeremyguez`, :user:`jeromekelleher`, :pr:`2246`, :pr:`2258`).
+
+- Add Colless tree imbalance index.
+  (:user:`jeremyguez`, :user:`jeromekelleher`, :issue:`2250`, :pr:`2266`, :pr:`2344`).
 
 **Breaking Changes**